Marge upstream 1d9233c3882afe888b9396f7f2aac917d4dcac4d (#21)
Fix `build_reconfigured_list` for ResourceImporterTexture.

+++ drivers/unix/os_unix.cpp
@@ -190,7 +190,7 @@ uint64_t OS_Unix::get_system_time_secs() const {
 uint64_t OS_Unix::get_system_time_msecs() const {
     struct timeval tv_now;
     gettimeofday(&tv_now, nullptr);
-    return uint64_t(tv_now.tv_sec * 1000 + tv_now.tv_usec / 1000);
+    return uint64_t(tv_now.tv_sec) * 1000 + uint64_t(tv_now.tv_usec) / 1000;
 }
 
 OS::Date OS_Unix::get_date(bool utc) const {
@@ -300,7 +300,7 @@ Error OS_Unix::execute(const String &p_path, const ListPOD<String> &p_arguments,
         }
         FILE *f = popen(qPrintable(argss.m_str), "r");
 
-		ERR_FAIL_COND_V_MSG(!f, ERR_CANT_OPEN, "Cannot pipe stream from process running with following arguments '" + argss + "'.")
+        ERR_FAIL_COND_V_MSG(!f, ERR_CANT_OPEN, "Cannot pipe stream from process running with following arguments '" + argss + "'.")
 
         char buf[65535];
 
@@ -316,7 +316,7 @@ Error OS_Unix::execute(const String &p_path, const ListPOD<String> &p_arguments,
         }
         int rv = pclose(f);
         if (r_exitcode)
-			*r_exitcode = WEXITSTATUS(rv);
+            *r_exitcode = WEXITSTATUS(rv);
 
         return OK;
     }
